+++ poem-openapi/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # Unreleased
 
 - impl `ParseFromParameter` for [T; const LEN: usize].
+- Add `example` attribute for `Object` macro.
 
 # [1.0.49] 2021-12-10